feat: add persistent storage claim
To persist data, this adds a persistent storage claim.

BREAKING CHANGES: The default storage path is now "/storage" to make
the configuration to use a PVC easier.

This requires action only if you have not explicitly set the storage path
AND if you require the storage to be at "/tmp".